Move client stuff to its own file
